feat: revert Enum ID to int and implement collision management tooling
[REVERT SUMMARY]
Reverted the internal ID system from 'long' back to 'int' to maintain compatibility with standard serialization and animator hashing. To mitigate the increased risk of ID collisions in the smaller 32-bit range, a comprehensive management and validation suite has been implemented.

[CHANGES]
* EnumAsset Core:
    - Restored _enumAssetId and instance values to 'int' type.
    - Adjusted bitmask and multiplier logic for integer constraints.

* Enum Asset Manager (New Tool):
    - Added project-wide scanner to detect prefix collisions between assets.
    - Implemented "Suggest Fix" to auto-assign available unique IDs.
    - Added "Manual Change" feature for testing and ID reorganization.
    - Integrated developer tooltips warning that ID updates require manual reference fixing in Scenes/Prefabs.

* EnumAsset Editor Enhancements:
    - Added [AUTO] vs [MANUAL] labeling for Asset IDs in the inspector.
    - Implemented high-visibility Designer Guidelines and ID persistence warnings.
    - Added a critical warning box urging a collision check before any asset modification.
    - Improved ReorderableList with color-coded duplicate detection for names and Local IDs.

[NOTE]
This change is a breaking change for existing serialized data stored as long. Ensure a full project refresh and collision scan is performed after merging.

+++ Assets/_Project/Scripts/Features/Actors/New/EntityStateManagement.cs
@@ -19,7 +19,7 @@ namespace Kope.Actor.New {
 		/// </remarks>
 		/// <param name="enumId">The unique identifier of the target state.</param>
 		/// <returns>The result of the transition attempt.</returns>
-		StateChangeResult ChangeState(long enumId, bool handleFallbackInternally = false);
+		StateChangeResult ChangeState(int enumId, bool handleFallbackInternally = false);
 
 		/// <summary>
 		/// The standard internal exit path for active states to return the entity to a neutral baseline.
@@ -53,12 +53,12 @@ namespace Kope.Actor.New {
 		private IAnimationComponentNew _animationComponent;
 
 		// EnumId of the idle state — stored separately so TransitionToIdle never needs a lookup key from outside.
-		private long _idleEnumId;
+		private int _idleEnumId;
 		private readonly EntityStateMachine _stateMachine = new();
 
 		// Primary lookup: enumId → logic SO.
 		// Null values are valid — purely data-driven states have no logic.
-		private readonly Dictionary<long, EntityStateBaseSO> _stateLookUp = new();
+		private readonly Dictionary<int, EntityStateBaseSO> _stateLookUp = new();
 
 		private void Awake() { if (this.loadOnStart) Init(); }
 
@@ -68,7 +68,7 @@ namespace Kope.Actor.New {
 
 			// --- Register idle first so it's always available as a fallback ---
 			var idleData = this.defaultIdleStateData.Profile.ToData();
-			long idleID = this.defaultIdleStateData.Profile.StatePicker.GetSelectedEnumId();
+			int idleID = this.defaultIdleStateData.Profile.StatePicker.GetSelectedEnumId();
 			this._idleEnumId = idleID;
 
 			EntityStateBaseSO idleLogic = null;
@@ -121,12 +121,11 @@ namespace Kope.Actor.New {
 				this._stateMachine.CurrentState.TickFixedUpdate();
 		}
 
-		public StateChangeResult ChangeState(long enumId, bool handleFallbackInternally = false) {
+		public StateChangeResult ChangeState(int enumId, bool handleFallbackInternally = false) {
 			// Attempt to find and schedule the transition.
 			var result = this._stateLookUp.TryGetValue(enumId, out var logic)
 				? this._stateMachine.ScheduleTransition(logic)
 				: StateChangeResult.Error_NotFound;
-
 
 			if (handleFallbackInternally && result != StateChangeResult.Success && result != StateChangeResult.Internal_Fallback) {
 				var stateName = logic.ProfileData.Name ?? $"EnumId {enumId}";
